Update grading_system.py
This python program implements a grading system using "if" and "elif" statements. It takes a student score as input and prints a corresponding grade based on the following scale:

. 70-100: Grade A
. 60-69: Grade B
. 50-59: Grade C
. 45-49: Grade D
. 40-45: Grade E
. 20-39: Grade F

For each valid score,the program also prints "YES" to confirm the grading was successful. If the score is outside the valid range, it prints "invalid score" and "NO"


Created by: Onyekachukwu Chimdinma Dabeluchukwu 
Reg No: 2024714003